+++ src/test/java/ar/edu/unrc/coeus/utils/StatisticCalculatorTest.java
@@ -16,22 +16,22 @@ class StatisticCalculatorTest {
             throws Exception {
         final StatisticCalculator estimator = new StatisticCalculator(5, 2);
         estimator.addSample(10.4);
-        estimator.addSample(4);
+        estimator.addSample(4.0);
         assertThat(estimator.getItemCounter(), is(2));
 
-        estimator.addSample(5);
-        estimator.addSample(9);
+        estimator.addSample(5.0);
+        estimator.addSample(9.0);
         assertThat(estimator.getItemCounter(), is(4));
 
-        estimator.addSample(2);
+        estimator.addSample(2.0);
         assertThat(estimator.getItemCounter(), is(5));
 
-        estimator.addSample(1);
+        estimator.addSample(1.0);
         assertThat(estimator.getItemCounter(), is(5));
-        estimator.addSample(1);
-        estimator.addSample(1);
-        estimator.addSample(1);
-        estimator.addSample(1);
+        estimator.addSample(1.0);
+        estimator.addSample(1.0);
+        estimator.addSample(1.0);
+        estimator.addSample(1.0);
         assertThat(estimator.getItemCounter(), is(5));
 
         assertThat(estimator.getAverage(), is(1.0d));
@@ -42,12 +42,12 @@ class StatisticCalculatorTest {
     void getAverage()
             throws Exception {
         final StatisticCalculator sc = new StatisticCalculator();
-        sc.addSample(10);
+        sc.addSample(10.0);
         sc.addSample(20.5);
         sc.addSample(2.698);
         sc.addSample(5.4678);
         sc.addSample(-26.5);
-        sc.addSample(0);
+        sc.addSample((double) 0);
         assertThat(sc.getAverage(), is(2.027633333333334d));
     }
 
@@ -60,23 +60,23 @@ class StatisticCalculatorTest {
         estimator.addSample(1.5);
         assertThat(estimator.printableFullCapacityAverage(), is("?"));
 
-        estimator.addSample(5);
-        estimator.addSample(9);
+        estimator.addSample(5.0);
+        estimator.addSample(9.0);
         assertThat(estimator.printableFullCapacityAverage(), is("?"));
 
-        estimator.addSample(2);
+        estimator.addSample(2.0);
         assertThat(estimator.printableFullCapacityAverage(), is("5,58"));
 
-        estimator.addSample(2);
-        estimator.addSample(2);
-        estimator.addSample(3);
-        estimator.addSample(2);
-        estimator.addSample(3);
+        estimator.addSample(2.0);
+        estimator.addSample(2.0);
+        estimator.addSample(3.0);
+        estimator.addSample(2.0);
+        estimator.addSample(3.0);
         assertThat(estimator.printableFullCapacityAverage(), is("2,4"));
 
         estimator = new StatisticCalculator(30, 2);
         for ( int i = 0; i < 29; i++ ) {
-            estimator.addSample(0);
+            estimator.addSample((double) 0);
         }
         estimator.addSample(3.2);
         assertThat(estimator.printableFullCapacityAverage(), is("0,11"));
@@ -88,9 +88,9 @@ class StatisticCalculatorTest {
             throws Exception {
         final StatisticCalculator estimator = new StatisticCalculator(5, 2);
         estimator.addSample(10.4);
-        estimator.addSample(4);
-        estimator.addSample(5);
-        estimator.addSample(9);
+        estimator.addSample(4.0);
+        estimator.addSample(5.0);
+        estimator.addSample(9.0);
         assertThat(estimator.getItemCounter(), is(4));
 
         estimator.reset();
